Rename schemas.py to devices.py
Rename file and update import
