+++ api/models/devices.py
